day14

+++ Arrow.h
@@ -14,7 +14,7 @@ public:
 	void dragDrop() override;
 	void use() override;
 	void init() override;
-	void shoot(Vector2 startPos, Vector2 shootVec, float damage) override;
+	void shoot(Vector2 startPos, Vector2 shootVec, int damage) override;
 
 	void setMoveVec(Vector2 _vec) { moveVec = _vec; }
 private:
